# By leanid (11) and others
# Via leanid (3) and Leanid Chaika (1)
* commit '99e493a642820957c5c9763a59e7f62f1ca6d7f0 [formerly 5cb968799fcd74ce1830d92fddf69cb23d6fcb59]': (21 commits)
  DF-6084 disable warnings as errors temporary continue in other task
  DF-6084 fix one more warnig
  DF-6084 fix ios build with cmake
  DF-6084 fix cmake flags
  DF-6084 start adding treat warnings as errors
  DF-6084 continue fix cmake treat warnings as error
  DF-6084 enable warnings as errors in all unit test with cmake
  DF-6084 fix review comments
  DF-6084 fix build resource editor
  DF-6084 fix win32 unit test
  DF-6084 fix compile, next need fix link
  DF-6084 continue
  DF-6084 continue fix warnings ios
  DF-6084 continue fix warnings
  DF-6084 resolve Save virtual name
  DF-6084 rename to resolve conflict MongoDB SaveObject and BaseObject SaveObject
  DF-6084 rename Load/Save BaseObject methods to fix warnings
  DF-6084 continue fix unittests warnings
  DF-6084 fix review comments. set warnings as errors on ios, android, win32
  DF-6084 fix review comments
  ...


Former-commit-id: 656cd3228f16ea50425a2d2a83cdc24361a7470d

+++ Sources/Internal/Render/Material/NMaterial.h
@@ -1085,7 +1085,7 @@ inline NMaterial* NMaterial::GetParent() const
 
 inline uint32 NMaterial::GetChildrenCount() const
 {
-    return children.size();
+    return static_cast<uint32>(children.size());
 }
 
 inline NMaterial* NMaterial::GetChild(uint32 index) const
